Crate sentry_slog

Sentry slog Integration.

The sentry slog integration consists of two parts, the SlogIntegration which configures how sentry should treat slog::Records, and the SentryDrain, which can be used to create a slog::Logger.

NOTE: This integration currently does not process any slog::KV pairs, but support for this will be added in the future.

Examples

use sentry::{init, ClientOptions};
use sentry_slog::{SentryDrain, SlogIntegration};

let integration = SlogIntegration::default();
let options = ClientOptions::default().add_integration(integration);
let _sentry = sentry::init(options);

let drain = SentryDrain::new(slog::Discard);
let root = slog::Logger::root(drain, slog::o!());

slog::info!(root, "recorded as breadcrumb");
slog::warn!(root, "recorded as regular event");

assert_eq!(
    captured_event.breadcrumbs.as_ref()[0].message.as_deref(),
    Some("recorded as breadcrumb")
);
assert_eq!(
    captured_event.message.as_deref(),
    Some("recorded as regular event")
);

slog::crit!(root, "recorded as exception event");

assert_eq!(captured_event.exception.len(), 1);

The integration can also be customized with a filter, and a mapper:

use sentry_slog::{exception_from_record, LevelFilter, RecordMapping, SlogIntegration};

let integration = SlogIntegration::default()
    .filter(|level| match level {
        slog::Level::Critical | slog::Level::Error => LevelFilter::Event,
        _ => LevelFilter::Ignore,
    })
    .mapper(|record, kv| RecordMapping::Event(exception_from_record(record, kv)));

Please not that the mapper can override any classification from the previous filter.

Structs

SentryDrain
A Drain which passes all Records to sentry.

SlogIntegration
The Sentry slog Integration.

Enums

LevelFilter
The Action that Sentry should perform for a slog::Level.

RecordMapping
Custom Mappers

Functions

breadcrumb_from_record
Creates a sentry Breadcrumb from the slog::Record.

convert_log_level
Converts a slog::Level to a sentry Level

default_filter
The default slog filter.

event_from_record
Creates a simple message Event from the slog::Record.

exception_from_record
Creates an exception Event from the slog::Record.